Adding automatization of change of inter version to a newer version
removing import keycloak realm automatically for old version of keycloak

+++ docker-stacks/02-keycloak/etc/hooks/after-start.d/10-configure-simva-realm.sh
@@ -1,22 +1,28 @@
 #!/usr/bin/env bash
 set -euo pipefail
 [[ "${DEBUG:-false}" == "true" ]] && set -x
-keycloakVersionFile="${SIMVA_CONFIG_HOME:-/home/vagrant/docker-stacks/config}/keycloak/keycloakversion.txt"
-if [[ ! -e "${SIMVA_PROJECT_DIR}/.simva-initialized" ]]; then 
-    $SIMVA_KEYCLOAK_VERSION > $keycloakVersionFile
-    docker exec -it 02-keycloak-keycloak-1 /opt/keycloak/bin/kc.sh import --file "/opt/keycloak/data/import/simva-realm-full.json" --override true --optimized
-else
-    if [[ -e "${SIMVA_CONFIG_HOME:-/home/vagrant/docker-stacks/config}/keycloak/simva-realm-export/${SIMVA_SSO_REALM:-simva}-realm-full.json" ]]; then
-        #previousVersion=0
-        #if [[ -e "${keycloakVersionFile}" ]]; then 
-        #    previousVersion=$(cat $keycloakVersionFile)    
-        #fi;
-        #if [[ ! $SIMVA_KEYCLOAK_VERSION == $previousVersion ]]; then
-        migrationinProgressFile="${SIMVA_CONFIG_HOME:-/home/vagrant/docker-stacks/config}/keycloak/simva-realm-export/.migrationinprogress"
-        if [[ -e "$migrationinProgressFile" ]]; then
-            docker exec -it 02-keycloak-keycloak-1 /opt/keycloak/bin/kc.sh import --file "/opt/keycloak/data/export/${SIMVA_SSO_REALM:-simva}-realm-full.json" --override true --optimized
-            #echo $SIMVA_KEYCLOAK_VERSION > $keycloakVersionFile
-            rm -f $migrationinProgressFile
+
+if [[ ${SIMVA_KEYCLOAK_VERSION%%.*} > 18 ]]; then 
+    if [[ ! -e "${SIMVA_PROJECT_DIR}/.simva-initialized" ]]; then 
+        echo "SIMVA is not initialized. Importing realm..." 
+        docker exec 02-keycloak-keycloak-1 /opt/keycloak/bin/kc.sh import --file "/opt/keycloak/data/import/simva-realm-full.json" --override true --optimized
+    else
+        echo "SIMVA is initialized." 
+        if [[ -e "${SIMVA_CONFIG_HOME:-/home/vagrant/docker-stacks/config}/keycloak/simva-realm-export/${SIMVA_SSO_REALM:-simva}-realm-full.json" ]]; then    
+            #previousVersion=0
+            #if [[ -e "${keycloakVersionFile}" ]]; then 
+            #    previousVersion=$(cat $keycloakVersionFile)    
+            #fi;
+            #if [[ ! $SIMVA_KEYCLOAK_VERSION == $previousVersion ]]; then
+            migrationinProgressFile="${SIMVA_CONFIG_HOME:-/home/vagrant/docker-stacks/config}/keycloak/simva-realm-export/.migrationinprogress"
+            if [[ -e "$migrationinProgressFile" ]]; then
+                echo "Migration in progress. Importing realm..."
+                docker exec 02-keycloak-keycloak-1 /opt/keycloak/bin/kc.sh import --file "/opt/keycloak/data/export/${SIMVA_SSO_REALM:-simva}-realm-full.json" --override true --optimized
+                #echo $SIMVA_KEYCLOAK_VERSION > $keycloakVersionFile
+                rm -f $migrationinProgressFile
+            fi;
         fi;
     fi;
-fi;
+else
+    echo "Please upgrade to a newer keycloak version ( > 18.*.*) before importing users and realm."
+fi;
